Added new entity "series", which refers to a group of conferences which belong together in some way. For the Tcl conferences going on in North-America, distinct from the conferences in Europe.
Extended the database and added a basic set of commands to create,
destroy and use series.

Extended the conference commands to deal with the new series.
This required an extension of the conference table schema.

The information about the series is available for insertion into
templates (@c:series@). The same is true for the detail information of
a series, the link to its index page (@c:series:link@).

At the moment this index page has to be maintained manually.  It is
planned to extend the website generator in the long-term future to
generate the sites of all known conferences, plus indices.

Beyond the above the contact commands got extensions as well. Creation
of a contact is now able to set tags, bio information, and related
contacts (affiliations, reps).