Extended database schema to record ordering for authors and speakers (Order matters in academic papers). Tables `submitter` and `talker`.
Modified all users of these tables to use the ordering information.
(Consolidated multiple identical sql command replicas into four helpers).

The default ordering is determined by the order of adding submitters
and speakers to a submission.

Currently the easiest way of reordering authors and/or speakers is to
drop the entire set and then add them back in the correct order.